Remove visibility sensor, simpplify unread counter, move NewsListItem into separate folder - css-modules example
