--- app/javascript/App/bundles/NewsfeedPage/NewsListItem.module.scss
@@ -1,20 +1,0 @@
-.newsListItem {
-  position: relative;
-}
-
-.cornerWrapper {
-  &::after {
-    width: 0;
-    height: 0;
-    border-top: 15px solid red;
-    border-right: 15px solid transparent;
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-  }
-}
-
-.title--hasRead {
-  color: #999;
-}
