+++ app/javascript/App/bundles/NewsfeedPage/NewsListItem/NewsListItem.module.scss
@@ -1,0 +1,3 @@
+.title--hasRead {
+  color: #999;
+}
